% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_gp_image}
\alias{create_gp_image}
\title{Calculates the Generalised Polarisation (GP) image of two given images.}
\usage{
create_gp_image(image_1, image_2, normalise = FALSE, display = FALSE)
}
\arguments{
\item{image_1}{A matrix representing the first image (preferably ordered channel).}

\item{image_2}{A matrix representing the second image (preferably disordered channel).}

\item{normalise}{Boolean value dictating whether GP values should be normalised between 0 and 1. Defaults to FALSE.}

\item{display}{Boolean value dictating whether GP image will be displayed. Defaults to FALSE.}
}
\value{
The GP image in matrix form.
}
\description{
Calculates the Generalised Polarisation (GP) image of two given images.
}
