% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{display_boundaries}
\alias{display_boundaries}
\title{Display all boundaries found on a given image.}
\usage{
display_boundaries(output, image_name, colour = c(255, 0, 0))
}
\arguments{
\item{output}{A string denoting the output file location. As defined in TOBLERONE, this is the location of the segmentation outputs.}

\item{image_name}{A string representing the name of the image.}

\item{colour}{A vector of length 3 representing the RGB colour code that defines the colour of the displayed loops. Defaults to red.}
}
\value{
Displays an image of the boundaries found from TOBLERONE.
}
\description{
This function plots the boundaries of all objects found from TOBLERONE.
}
