% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{display_segmentation}
\alias{display_segmentation}
\title{Display a segmentation result from TOBLERONE.}
\usage{
display_segmentation(
  output,
  image_name,
  stage = 1,
  colourblind_palette = FALSE
)
}
\arguments{
\item{output}{A string denoting the output file location. As defined in TOBLERONE, this is the location of the segmentation outputs.}

\item{image_name}{A string representing the name of the image.}

\item{stage}{Set equal to 1 for primary segmentation and 2 for secondary segmentation. Defaults to 1.}
}
\value{
Displays an image of the whole segmentation.
}
\description{
This function plots the objects found from either the primary or
secondary segmentation.
}
