% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_gp_values}
\alias{get_gp_values}
\title{Calculates the Generalised Polarisation (GP) image of returns the GP values along the line profiles identified by TOBLERONE.}
\usage{
get_gp_values(image_1, image_2, image_name, output)
}
\arguments{
\item{image_1}{A matrix representing the first image (preferably ordered channel).}

\item{image_2}{A matrix representing the second image (preferably disordered channel).}

\item{image_name}{A string representing the name of the image originally used in TOBLERONE.}

\item{output}{A string denoting the output file location as used in TOBLERONE originally.}
}
\value{
A list of the GP values along the line profiles identified by TOBLERONE.
}
\description{
Calculates the Generalised Polarisation (GP) image of returns the GP values along the line profiles identified by TOBLERONE.
}
