% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{toblerone}
\alias{toblerone}
\title{Perform TOBLERONE on a given image.}
\usage{
toblerone(
  image,
  image_name,
  output,
  failsafe = TRUE,
  use_image_decomposition = TRUE,
  check_objects = TRUE,
  base_window_size = 1,
  identify_boundary = TRUE,
  smooth_image = TRUE,
  colourblind_palette = FALSE
)
}
\arguments{
\item{image}{A matrix representing the image to be analysed.}

\item{image_name}{A string representing the name of the image, of which all files will be saved under.}

\item{output}{A string denoting the output file location. All results and calculations will be output here.}
}
\value{
A set of images and "loop" csv files corresponding to the pixel coordinates of the boundaries found by TOBLERONE.
}
\description{
This function performs TOBLERONE on a given grayscale image. It is recommended
that the image is represented in matrix form with all values between 0 and 1.
}
