% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{toblerone3D}
\alias{toblerone3D}
\title{Perform 3D TOBLERONE on a given stack.}
\usage{
toblerone3D(
  stack,
  stack_name,
  output,
  use_image_decomposition = TRUE,
  brightness_scale = 1,
  colourblind_palette = FALSE,
  output_stack = TRUE,
  persistence = NULL,
  repeat_thresholding = TRUE
)
}
\arguments{
\item{stack}{An array representing the stack of images to be analysed.}

\item{stack_name}{A character representing the name of the stack, of which all files will be saved under.}

\item{output}{A character denoting the output file location. All results and calculations will be output here.}

\item{use_image_decomposition}{Logical. Dictates whether to remove pixels with 0 intensity.}

\item{brightness_scale}{Optional numeric which scales the brightness intensity of each pixel in the image. Intensities over the maximum are returned to 1.}

\item{colourblind_palette}{Logical. Dictates whether to use a colourblind-friendly colour palette on masks.}

\item{output_stack}{Logical. Dictates whether to output segmented image as stack (otherwise just outputs summary statistics).}

\item{persistence}{Numeric for absolute persistence threshold. If left as NULL, user will be consulted during runtime.}

\item{repeat_thresholding}{Logical. If TRUE, user is prompted to input persistence threshold until desired segmentation is reached.}
}
\value{
A list of images corresponding to the voxel coordinates of the segmentation found by TOBLERONE.
}
\description{
This function performs TOBLERONE on a given stack of grayscale images. It is recommended
that the image is represented in matrix form with all values between 0 and 1.
}
