% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{toblerone3D}
\alias{toblerone3D}
\title{Perform 3D TOBLERONE on a given stack.}
\usage{
toblerone(
  stack,
  stack_name,
  output,
  use_image_decomposition = TRUE,
    brightness_scale = 1,
  colourblind_palette = FALSE
)
}
\arguments{
\item{stack}{An array representing the stack of images to be analysed.}

\item{stack_name}{A string representing the name of the stack, of which all files will be saved under.}

\item{output}{A string denoting the output file location. All results and calculations will be output here.}
}
\value{
A set of images corresponding to the pixel coordinates of the segmentation found by TOBLERONE.
}
\description{
This function performs TOBLERONE on a given stack of grayscale images. It is recommended
that the image is represented in matrix form with all values between 0 and 1.
}
